requirements.txt
fastapi
uvicorn
sqlalchemy
pymysql
pdfminer.six
docx2txt
spacy
python-dotenv